refactor: bottom navigation bar as a shared widget
