+++ lib/presenter/home_page/home_page.dart
@@ -1,48 +1,50 @@
-import 'package:crypto/presenter/home_page/widgets/bottom_navigationbar.dart';
-import 'package:crypto/presenter/home_page/widgets/card_coin.dart';
-import 'package:crypto/presenter/home_page/widgets/upper_card_crypto.dart';
+import 'package:crypto/presenter/movements/movements_page.dart';
+import 'package:crypto/presenter/portifolio_page/portifolio_page.dart';
 import 'package:flutter/material.dart';
 
-class HomePage extends StatelessWidget {
-  const HomePage({
-    Key? key,
-  }) : super(key: key);
+import '../../shared/colors.dart';
+
+class HomePage extends StatefulWidget {
+  const HomePage({super.key});
+
+  @override
+  State<HomePage> createState() => _HomePageState();
+}
+
+class _HomePageState extends State<HomePage> {
+  int _selectedIndex = 0;
+
+  static const List<Widget> _widgetOptions = <Widget>[
+    PortifolioPage(),
+    MovementsPage(),
+  ];
+
+  void _onItemTapped(int index) {
+    setState(() {
+      _selectedIndex = index;
+    });
+  }
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      body: SingleChildScrollView(
-        child: Column(
-          children: const [
-            UpperCardCrypto(),
-            Divider(thickness: 1),
-            CardCoin(
-              abbreviation: 'BTC',
-              name: 'Bitcoin',
-              price: 6557.00,
-              variation: 0.65,
-              iconImage: 'assets/icons/bitecoin.png',
-            ),
-            Divider(thickness: 1),
-            CardCoin(
-              abbreviation: 'ETH',
-              name: 'Ethereum',
-              price: 7996.00,
-              variation: 0.94,
-              iconImage: 'assets/icons/ethereum.png',
-            ),
-            Divider(thickness: 1),
-            CardCoin(
-              abbreviation: 'LTC',
-              name: 'Litecoin',
-              price: 245.00,
-              variation: 0.82,
-              iconImage: 'assets/icons/litecoin.png',
-            ),
-          ],
-        ),
+      backgroundColor: Colors.grey,
+      body: _widgetOptions.elementAt(_selectedIndex),
+      bottomNavigationBar: BottomNavigationBar(
+        items: const <BottomNavigationBarItem>[
+          BottomNavigationBarItem(
+              icon: ImageIcon(
+                AssetImage('assets/icons/icon_warren.png'),
+              ),
+              label: 'Portifólio'),
+          BottomNavigationBarItem(
+              icon: ImageIcon(AssetImage('assets/icons/cent.png')),
+              label: 'Movimentações'),
+        ],
+        currentIndex: _selectedIndex,
+        selectedItemColor: magenta,
+        onTap: _onItemTapped,
       ),
-      bottomNavigationBar: const BottomNavigationBarApp(),
     );
   }
 }
