+++ lib/presenter/movements/movements_page.dart
@@ -1,0 +1,25 @@
+import 'package:flutter/material.dart';
+
+class MovementsPage extends StatefulWidget {
+  const MovementsPage({super.key});
+
+  @override
+  State<MovementsPage> createState() => _MovementsPageState();
+}
+
+class _MovementsPageState extends State<MovementsPage> {
+  @override
+  Widget build(BuildContext context) {
+    return SafeArea(
+      child: Scaffold(
+        body: Column(
+          children: const [
+            Text(
+              'olaaaaa',
+            )
+          ],
+        ),
+      ),
+    );
+  }
+}
